requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.54.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
faiss-cpu==1.7.4
pandas==2.1.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0
aiofiles==23.2.0
python-multipart==0.0.6
jinja2==3.1.2
gspread==6.2.1
google-auth==2.40.3
